requirements.txt
streamlit
langchain-community
langchain-core
langchain_openai 
python-dotenv
PyPDF2
faiss-cpu
sentence-transformers